🚚 rename: 파일, 폴더명, emotion 선언 위치 수정
