+++ components/Gnb/components/NotiButton.tsx
@@ -2,21 +2,6 @@ import styled from "@emotion/styled";
 import Image from "next/image";
 import { useState } from "react";
 
-const ModalContainer = styled.div`
-  position: absolute;
-  z-index: 1000;
-  top: +50px;
-  right: 350px;
-  height: 200px;
-  width: 368px;
-  padding: 0 20px 24px;
-  overflow-y: auto;
-  border: 1px solid var(--The-julge-gray-30);
-  border-radius: 10px;
-  background-color: var(--The-julge-purple-10);
-  box-shadow: 0 2px 8px 0 rgb(120 116 134 / 25%);
-`;
-
 export default function NotiButton({ activeStatus }: { activeStatus: string }) {
   const [isModalOpen, setIsModalOpen] = useState(false);
 
@@ -42,3 +27,18 @@ export default function NotiButton({ activeStatus }: { activeStatus: string }) {
     </>
   );
 }
+
+const ModalContainer = styled.div`
+  position: absolute;
+  z-index: 1000;
+  top: +50px;
+  right: 350px;
+  height: 200px;
+  width: 368px;
+  padding: 0 20px 24px;
+  overflow-y: auto;
+  border: 1px solid var(--The-julge-gray-30);
+  border-radius: 10px;
+  background-color: var(--The-julge-purple-10);
+  box-shadow: 0 2px 8px 0 rgb(120 116 134 / 25%);
+`;
