+++ components/Gnb/styles/textStyles.ts
